Feat: MainActivity.kt 통합용으로 수정
